my homework
created a class for the master token, initialized with the master credentials, and replaced the credentials string in the AuthViewController with an instance of the class. 

I'm not sure how these two documents are supposed to talk with each other without an "import" at the top, but it seems none of the other swift files do that so I assume they're connected through file path.